修复所有编译错误并成功安装Python模块
修复内容：
1. state.rs: 使用 NUM_PLAYERS 常量替代硬编码的 4
2. game_state.rs: 添加 Tile 类型导入
3. tensor.rs: 修复类型不匹配（u8 vs usize）

安装步骤：
1. 安装 maturin: pip install maturin
2. 编译并安装: python3 -m maturin develop
3. 验证导入成功

结果：所有模块可以成功导入，系统已准备好进行训练

+++ rust/src/python/tensor.rs
@@ -137,7 +137,7 @@ pub fn state_to_tensor(
             // 获取该玩家的弃牌（从弃牌历史中筛选，保持原始顺序）
             let player_discards: Vec<_> = game_state.discard_history
                 .iter()
-                .filter(|record| record.player_id == p_id)
+                .filter(|record| record.player_id as usize == p_id)
                 .collect();
             
             // 为该玩家分配 4 个平面（Plane 14 + p_id * 4 到 Plane 17 + p_id * 4）
